Relocate extraction procedures to Code module
