+++ core-strato/ethereum-vm/bench/CodeBench.hs
@@ -4,23 +4,10 @@
 {-# OPTIONS_GHC -fno-warn-missing-fields #-}
 import Control.Monad
 import Criterion.Main
-import Data.Bits
 import qualified Data.ByteString        as B
-import qualified Data.ByteString.Unsafe as BU
-import Data.Primitive.ByteArray
-import Foreign.Ptr
-import Foreign.Storable
-import GHC.Exts
-import GHC.Integer.GMP.Internals (Integer(..), BigNat(..))
-import GHC.Word
-import System.Endian
-import System.IO.Unsafe
-
-import Network.Haskoin.Internals (BigWord(..))
 
 import Blockchain.Data.Code
-import Blockchain.Strato.Model.ExtendedWord
-import Blockchain.Util
+import Blockchain.VM.Code
 
 import Test.Hspec
 
@@ -28,55 +15,6 @@ import Test.Hspec
 exampleCode :: Code
 exampleCode = Code $ B.pack $ [0..255]
 
--- Unoptimized push, for 8-24 bytes that are too infrequently seen
--- to bother writing a specialization.
-defaultExtract :: Code -> Int -> Int -> Word256
--- TODO(tim): Use fastBytesToWord256 once available
-defaultExtract (Code bs) off len = fromIntegral
-                                 . bytes2Integer
-                                 . B.unpack
-                                 . B.take len
-                                 . B.drop off
-                                 $ bs
-defaultExtract _ _ _ = error "precompiled contracts cannot slice code"
-
-fastExtractByte :: Code -> Int -> Word256
-fastExtractByte (Code !code) !off = let !(W8# b#) = BU.unsafeIndex code off
-                                 in BigWord (S# (word2Int# b#))
-fastExtractByte _ _ = error "cannot slice out of precompiled"
-
--- Used to push 2-7 bytes
-fastExtractSingle :: Code -> Int -> Int -> Word256
-fastExtractSingle (Code !code) !off !len = unsafePerformIO . BU.unsafeUseAsCString code $ \ptr -> do
-  let !offPtr = castPtr ptr :: Ptr Word64
-      !delta = 64 - (8 * len)
-  -- This may read past the end of the bytestring, but if the read is allowed
-  -- those garbage bytes are truncated by the shift.
-  !rawBits <- peekByteOff offPtr off
-  let !(W64# w#) = toBE64 rawBits `shiftR` delta
-  return $! BigWord (S# (word2Int# w#))
-fastExtractSingle _ _ _ = error "cannot slice out of precompiled"
-
--- Used to push 25-32 bytes
-fastExtractQuad :: Code -> Int -> Int -> Word256
-fastExtractQuad (Code !code) !off !len = unsafePerformIO . BU.unsafeUseAsCString code $ \ptr -> do
-  let !offPtr = castPtr (plusPtr ptr (off + len)) :: Ptr Word64
-  dst <- newByteArray 32
-  ll <- peekElemOff offPtr (-1)
-  lh <- peekElemOff offPtr (-2)
-  hl <- peekElemOff offPtr (-3)
-  -- This might be a violation: we read before the beginning of the bytestring.
-  -- However if the read is allowed, the garbage bytes are masked off.
-  hh <- peekElemOff offPtr (-4)
-
-  writeByteArray dst 0 $! toBE64 ll
-  writeByteArray dst 1 $! toBE64 lh
-  writeByteArray dst 2 $! toBE64 hl
-  let !mask = bit (8 * (len - 24)) - 1
-  writeByteArray dst 3 $! toBE64 hh .&. mask
-  !(ByteArray ba#) <- unsafeFreezeByteArray dst
-  return (BigWord (Jp# (BN# ba#)))
-fastExtractQuad _ _ _ = error "cannot slice out of precompiled"
 
 benchExtract1Slow :: Benchmark
 benchExtract1Slow = bench "extract1 slow"
