+++ core-strato/ethereum-vm/src/Blockchain/VM/Code.hs
@@ -1,9 +1,20 @@
 {-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE MagicHash    #-}
 module Blockchain.VM.Code where
 
+import           Data.Bits
 import qualified Data.ByteString              as B
+import qualified Data.ByteString.Unsafe       as BU
 import qualified Data.IntSet                  as I
+import           Data.Primitive.ByteArray
+import           Foreign.Ptr
+import           Foreign.Storable
+import           GHC.Exts
+import           GHC.Integer.GMP.Internals (Integer(..), BigNat(..))
+import           GHC.Word
 import           Numeric
+import           System.Endian
+import           System.IO.Unsafe
 import           Text.PrettyPrint.ANSI.Leijen
 
 import qualified Blockchain.Colors            as CL
@@ -11,6 +22,7 @@ import           Blockchain.Data.Code
 import           Blockchain.Format
 import           Blockchain.Util
 import           Blockchain.VM.Opcodes
+import           Network.Haskoin.Internals (BigWord(..), Word256)
 
 
 getOperationAt::Code->CodePointer->(Operation, CodePointer)
@@ -51,3 +63,53 @@ compile::[Operation]->Code
 compile x = Code bytes
   where
     bytes = B.pack $ op2OpCode =<< x
+
+-- Unoptimized push, for 8-24 bytes that are too infrequently seen
+-- to bother writing a specialization.
+defaultExtract :: Code -> Int -> Int -> Word256
+-- TODO(tim): Use fastBytesToWord256 once available
+defaultExtract (Code bs) off len = fromIntegral
+                                 . bytes2Integer
+                                 . B.unpack
+                                 . B.take len
+                                 . B.drop off
+                                 $ bs
+defaultExtract _ _ _ = error "precompiled contracts cannot slice code"
+
+fastExtractByte :: Code -> Int -> Word256
+fastExtractByte (Code !code) !off = let !(W8# b#) = BU.unsafeIndex code off
+                                    in BigWord (S# (word2Int# b#))
+fastExtractByte _ _ = error "cannot slice out of precompiled"
+
+-- Used to push 2-7 bytes
+fastExtractSingle :: Code -> Int -> Int -> Word256
+fastExtractSingle (Code !code) !off !len = unsafePerformIO . BU.unsafeUseAsCString code $ \ptr -> do
+  let !offPtr = castPtr ptr :: Ptr Word64
+      !delta = 64 - (8 * len)
+  -- This may read past the end of the bytestring, but if the read is allowed
+  -- those garbage bytes are truncated by the shift.
+  !rawBits <- peekByteOff offPtr off
+  let !(W64# w#) = toBE64 rawBits `shiftR` delta
+  return $! BigWord (S# (word2Int# w#))
+fastExtractSingle _ _ _ = error "cannot slice out of precompiled"
+
+-- Used to push 25-32 bytes
+fastExtractQuad :: Code -> Int -> Int -> Word256
+fastExtractQuad (Code !code) !off !len = unsafePerformIO . BU.unsafeUseAsCString code $ \ptr -> do
+  let !offPtr = castPtr (plusPtr ptr (off + len)) :: Ptr Word64
+  dst <- newByteArray 32
+  ll <- peekElemOff offPtr (-1)
+  lh <- peekElemOff offPtr (-2)
+  hl <- peekElemOff offPtr (-3)
+  -- This might be a violation: we read before the beginning of the bytestring.
+  -- However if the read is allowed, the garbage bytes are masked off.
+  hh <- peekElemOff offPtr (-4)
+
+  writeByteArray dst 0 $! toBE64 ll
+  writeByteArray dst 1 $! toBE64 lh
+  writeByteArray dst 2 $! toBE64 hl
+  let !mask = bit (8 * (len - 24)) - 1
+  writeByteArray dst 3 $! toBE64 hh .&. mask
+  !(ByteArray ba#) <- unsafeFreezeByteArray dst
+  return (BigWord (Jp# (BN# ba#)))
+fastExtractQuad _ _ _ = error "cannot slice out of precompiled"
